requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
aiofiles
pytesseract
pdfplumber
Pillow
python-multipart